Move copyGoogleServices to plugin
